Reduce precision
Don't need Int64 to store binary data.

+++ src/processing/reads.jl
@@ -122,7 +122,7 @@ function process_exclusions(
             end
 
             exclusion_zones_bool = create_exclusion_zones(bathy_subset, vessel_draft)
-            write(exclusion_tif_path, convert.(Int64, exclusion_zones_bool); force=true)
+            write(exclusion_tif_path, convert.(Int8, exclusion_zones_bool); force=true)
         end
 
         exclusion_zones_df = polygonize_binary(exclusion_zones_bool)
